+++ ebcli/display/environments.py
@@ -17,9 +17,8 @@ import time
 
 from ebcli.lib.utils import get_local_time
 from ebcli.resources.strings import prompts, responses
-
-from . import term
-from ..core import io
+from ebcli.display import term
+from ebcli.core import io
 
 locale.setlocale(locale.LC_ALL, 'C')
 from datetime import datetime, timedelta
@@ -28,7 +27,7 @@ from botocore.compat import six
 from ebcli.display.data_poller import format_time_since, DataPoller
 from ebcli.display.screen import Screen
 from ebcli.lib import utils
-from ebcli.objects.exceptions import ValidationError, NotFoundError
+from ebcli.objects.exceptions import NotFoundError, ValidationError
 from ebcli.resources.strings import strings
 
 Queue = six.moves.queue.Queue