+++ ebcli/operations/logsops.py
@@ -11,7 +11,7 @@
 # ANY KIND, either express or implied. See the License for the specific
 # language governing permissions and limitations under the License.
 
-from datetime import datetime, timedelta
+from datetime import datetime
 import os
 import sys
 import threading
@@ -20,17 +20,21 @@ import time
 from cement.utils.misc import minimal_logger
 from six import iteritems
 
-from ..core import fileoperations, io
-from ..lib import elasticbeanstalk, utils, cloudwatch
-from ..lib.aws import MaxRetriesError
-from ..resources.strings import strings, prompts
-from ..objects.exceptions import ServiceError
-from . import commonops
+from ebcli.core import fileoperations, io
+from ebcli.lib import elasticbeanstalk, utils, cloudwatch
+from ebcli.lib.aws import MaxRetriesError
+from ebcli.resources.strings import strings, prompts
+from ebcli.resources.statics import namespaces, option_names
+from ebcli.objects.exceptions import ServiceError, NotFoundError
+from ebcli.operations import commonops
 
 LOG = minimal_logger(__name__)
+TAIL_LOG_SIZE = 100
+DEFAULT_LOG_STREAMING_PATH = 'var/log/eb-activity.log'
+BEANSTALK_LOG_PREFIX = '/aws/elasticbeanstalk'
 
 
-def logs(env_name, info_type, do_zip=False, instance_id=None):
+def retrieve_beanstalk_logs(env_name, info_type, do_zip=False, instance_id=None):
     # Request info
     result = elasticbeanstalk.request_environment_info(env_name, info_type)
 
@@ -49,7 +53,7 @@ def get_logs(env_name, info_type, do_zip=False, instance_id=None):
     result = elasticbeanstalk.retrieve_environment_info(env_name, info_type)
 
     """
-    Results are ordered with latest last, we just want the latest
+        Results are ordered with latest last, we just want the latest
     """
     log_list = {}
     for log in result['EnvironmentInfo']:
@@ -58,7 +62,10 @@ def get_logs(env_name, info_type, do_zip=False, instance_id=None):
         log_list[i_id] = url
 
     if instance_id:
-        log_list = {instance_id: log_list[instance_id]}
+        try:
+            log_list = {instance_id: log_list[instance_id]}
+        except KeyError:
+            raise NotFoundError(strings['beanstalk-logs.badinstance'].replace('{instance_id}', instance_id))
 
     if info_type == 'bundle':
         # save file, unzip, place in logs directory
@@ -111,22 +118,38 @@ def get_logs(env_name, info_type, do_zip=False, instance_id=None):
         io.echo_with_pager(os.linesep.join(data))
 
 
-def stream_logs(env_name, sleep_time=2):
-    log_group_name = 'awseb-' + env_name + '-activity'
+def stream_cloudwatch_logs(env_name, sleep_time=2, log_group=None, instance_id=None):
+    """
+        This function will stream logs to the terminal for the log group given, if multiple streams are found we will
+        spawn multiple threads with each stream switch between them to stream all streams at the same time.
+        :param env_name: environment name
+        :param sleep_time: sleep time to refresh the logs from cloudwatch
+        :param log_group: cloudwatch log group
+        :param instance_id: since all of our log streams are instance ids we require this if we want a single stream
+    """
+    if log_group is None:
+        log_group = 'awseb-{0}-activity'.format(env_name)
+        log_name = 'eb-activity.log'
+    else:
+        log_name = get_log_name(log_group)
     stream_names = []
-
     streamer = io.get_event_streamer()
-    streamer.prompt = ' -- eb-activity log -- (Ctrl+C to exit)'
+    streamer.prompt = ' -- {0} -- (Ctrl+C to exit)'.format(log_name)
     jobs = []
     while True:
-        new_names = cloudwatch.get_all_stream_names(log_group_name)
+        try:
+            new_names = cloudwatch.get_all_stream_names(log_group, instance_id)
+        except:
+            raise NotFoundError(strings['cloudwatch-stream.notsetup'])
+        if len(new_names) == 0:
+            raise NotFoundError(strings['cloudwatch-logs.nostreams'].replace('{log_group}', log_group))
         for name in new_names:
             if name not in stream_names:
                 stream_names.append(name)
 
                 p = threading.Thread(
                     target=_stream_single_stream,
-                    args=(log_group_name, name, streamer, sleep_time))
+                    args=(log_group, name, streamer, sleep_time))
                 p.daemon = True
                 jobs.append(p)
                 p.start()
@@ -168,3 +191,182 @@ def _stream_single_stream(log_group_name, stream_name, streamer, sleep_time=4):
             # Various things
             LOG.debug('Exception raised: ' + str(e))
             # Loop will cause a retry
+
+
+def retrieve_cloudwatch_logs(log_group, info_type, do_zip=False, instance_id=None):
+    # Get the log streams, a.k.a. the instance ids in the log group
+    """
+        Retrieves cloudwatch logs for every stream under the log group unless the instance_id is specified. If tail
+         logs is enabled we will only get the last 100 lines and return the result to a pager for the user to use. If
+         bundle info type is chosen we will get all of the logs and save them to a dir under .elasticbeanstalk/logs/
+        and if zip is enabled we will zip those logs for the user.
+        :param log_group: cloudwatch log group
+        :param info_type: can be 'tail' or 'bundle'
+        :param do_zip: boolean to determine if we should zip the logs we retrieve
+        :param instance_id: if we only want a single instance we can specify it here
+    """
+    log_streams = cloudwatch.describe_log_streams(log_group, log_stream_name_prefix=instance_id)
+    instance_ids = []
+
+    if len(log_streams['logStreams']) == 0:
+        io.log_error(strings['logs.nostreams'])
+
+    for stream in log_streams['logStreams']:
+        instance_ids.append(stream['logStreamName'])
+
+    # This is analogous to getting the full logs
+    if info_type == 'bundle':
+        # Create directory to store logs
+        logs_folder_name = datetime.now().strftime("%y%m%d_%H%M%S")
+        logs_location = fileoperations.get_logs_location(logs_folder_name)
+        os.makedirs(logs_location)
+        # Get logs for each instance
+        for instance_id in instance_ids:
+            full_logs = get_cloudwatch_stream_logs(log_group, instance_id)
+            full_filepath = '{0}/{1}.log'.format(logs_location, instance_id)
+            log_file = open(full_filepath, 'w+')
+            log_file.write(full_logs)
+            log_file.close()
+            fileoperations.set_user_only_permissions(full_filepath)
+
+        if do_zip:
+            fileoperations.zip_up_folder(logs_location, logs_location + '.zip')
+            fileoperations.delete_directory(logs_location)
+
+            logs_location += '.zip'
+            fileoperations.set_user_only_permissions(logs_location)
+            io.echo(strings['logs.location'].replace('{location}', logs_location))
+        else:
+            io.echo(strings['logs.location'].replace('{location}', logs_location))
+            # create symlink to logs/latest
+            latest_location = fileoperations.get_logs_location('latest')
+            try:
+                os.unlink(latest_location)
+            except OSError:
+                # doesn't exist. Ignore
+                pass
+            try:
+                os.symlink(logs_location, latest_location)
+                io.echo('Updated symlink at', latest_location)
+            except OSError:
+                # Oh well.. we tried.
+                ## Probably on windows, or logs/latest is not a symlink
+                pass
+
+    else:
+        # print logs
+        all_logs = ""
+        for instance_id in instance_ids:
+            tail_logs = get_cloudwatch_stream_logs(log_group, instance_id, num_log_events=TAIL_LOG_SIZE)
+            all_logs += '\n\n============= {0} - {1} ==============\n\n'.format(str(instance_id), get_log_name(log_group))
+            all_logs += tail_logs
+        io.echo_with_pager(all_logs)
+
+
+def get_cloudwatch_stream_logs(log_group_name, stream_name, num_log_events=None):
+    """
+        Will get logs events from cloudwatch and append them to a single string to output with each line prefixed with
+         the stream name.
+        :param log_group_name: cloudwatch log group
+        :param stream_name: cloudwatch stream name
+        :param num_log_events: number of log events to retrieve; default is cloudwatch's max: 10k or 1MB of messages
+        :return: single string will all log events concatenated together
+    """
+    full_log = ''
+    try:
+        response = cloudwatch.get_log_events(log_group_name, stream_name, limit=num_log_events)
+
+        if response and response.get('events'):
+            for event in response.get('events'):
+                message = event.get('message')
+
+                full_log += '[{}] {}\n'.format(stream_name, message)
+
+    except ServiceError as e:
+        # Something went wrong getting the stream
+        # It probably doesnt exist anymore.
+        LOG.debug('Received service error {}'.format(e))
+        return
+    except Exception as e:
+        # We want to swallow all exceptions or else they will be
+        # printed as a stack trace to the Console
+        # Exceptions are typically connections reset and
+        # Various things
+        LOG.debug('Exception raised: ' + str(e))
+        # Loop will cause a retry
+    return full_log
+
+
+def log_streaming_enabled(app_name, env_name):
+    """
+        Checks if log streaming is enabled for the given environment
+        :param app_name: application name
+        :param env_name: environment name
+        :return: boolean if the given environment has log stremaing enabled
+    """
+    config_settings = elasticbeanstalk.describe_configuration_settings(app_name, env_name)
+    stream_enabled = elasticbeanstalk.get_specific_configuration(config_settings, namespaces.CLOUDWATCH_LOGS,
+                                                                 option_names.STREAM_LOGS)
+    if stream_enabled is not None and stream_enabled == 'true':
+        return True
+    return False
+
+
+def disable_cloudwatch_logs(env_name):
+    """
+        Disables cloudwatch log streaming for the given environment
+        :param env_name: environment name
+    """
+    # Add option settings needed for log streaming
+    option_settings = [
+        elasticbeanstalk.create_option_setting(
+            namespaces.CLOUDWATCH_LOGS,
+            option_names.STREAM_LOGS,
+            'false'),
+    ]
+    io.echo(strings['cloudwatch-logs.disable'])
+    commonops.update_environment(env_name, changes=option_settings, nohang=False)
+
+
+def enable_cloudwatch_logs(env_name):
+    # Add option settings needed for log streaming
+    """
+        Enables cloudwatch log streaming for the given environment
+        :param env_name: environment name
+    """
+    option_settings = [
+        elasticbeanstalk.create_option_setting(
+            namespaces.CLOUDWATCH_LOGS,
+            option_names.STREAM_LOGS,
+            'true'
+        ),
+    ]
+    io.echo(strings['cloudwatch-logs.enable'])
+    io.echo(strings['cloudwatch-logs.link'].replace('{region}', commonops.get_default_region()).replace('{env_name}', env_name))
+    commonops.update_environment(env_name, changes=option_settings, nohang=False)
+
+
+# TODO: Maybe not foce all log groups to be beanstalk specific
+def beanstalk_log_group_builder(env_name, filepath=DEFAULT_LOG_STREAMING_PATH):
+    """
+        The log builder will take an optional filepath and attempt to build the log group specifically for groups created
+         by the Elastic Beanstalk service.
+        :param env_name: current environment being used
+        :param filepath: path that is apart of the log_stream
+        :return: the full log group for a beanstalk log group
+    """
+    if filepath is None:
+        filepath = DEFAULT_LOG_STREAMING_PATH
+    elif filepath.startswith('{0}/{1}'.format(BEANSTALK_LOG_PREFIX, env_name)):
+        return filepath
+    return '{0}/{1}/{2}'.format(BEANSTALK_LOG_PREFIX, env_name, filepath)
+
+
+def get_log_name(log_group):
+    """
+        This will parse the log group to get the filename of the log. Benastalk creates log groups with the the filepath of
+         the log, example: '/aws/elasticbeanstalk/env-name/var/log/eb-activity.log'.
+        :param log_group: full or partial log group
+        :return: the last string on the path, i.e. the filename
+    """
+    return log_group.split('/')[-1]